+++ web/documentserver-example/php/assets/js/jscript.js
@@ -325,6 +325,22 @@ if (typeof jQuery != "undefined") {
         });
     });
 
+    jq(document).on("click", ".clear-all", function () {
+        if (confirm("Delete all the files?")) {
+            jq.ajax({
+                async: true,
+                contentType: "text/xml",
+                type: "delete",
+                url: "delete",
+                complete: function (data) {
+                    if (JSON.parse(data.responseText).status == 'success') {
+                        window.location.reload(true);
+                    }
+                }
+            });
+        }
+    });
+
     jq(document).on("click", "#createSample", function () {
         jq(".try-editor").each(function () {
             var href = jq(this).attr("href");