+++ web/documentserver-example/ruby/app/assets/javascripts/jscript.js
@@ -307,11 +307,28 @@ if (typeof jQuery != "undefined") {
         jq.ajax({
             async: true,
             contentType: "text/xml",
+            type: "delete",
             url: requestAddress,
             complete: function (data) {
                 document.location.reload();
             }
         });
+    });
+
+    jq(document).on("click", ".clear-all", function () {
+        if (confirm("Delete all the files?")) {
+            jq.ajax({
+                async: true,
+                contentType: "text/xml",
+                type: "delete",
+                url: "remove",
+                complete: function (data) {
+                    if (JSON.parse(data.responseText).success) {
+                        window.location.reload(true);
+                    }
+                }
+            });
+        }
     });
 
         function showUserTooltip (isMobile) {